+++ storage/schemas.py
@@ -1,3 +1,5 @@
 from storage.base import Base
 from storage.schema_version import SchemaVersion
 from storage.triplets import Triplets
+from storage.references import References
+from storage.structures import Structures