+++ update/common.py
@@ -7,7 +7,9 @@ QLEVER_URL = "https://qlever.cs.uni-freiburg.de/api/wikidata"
 
 
 def sparql_to_csv(query: str, url: str = WD_URL) -> str:
-    return get(url, params={"query": query}, headers={"Accept": "text/csv"}).text
+    return get(url, params={"query": query}, headers={"Accept": "text/csv",
+                                                      "Accept-Encoding": "gzip,deflate",
+                                                      "User-Agent": "LOTUS project database dumper"}).text
 
 
 def remove_wd_entity_prefix(text: str) -> str: