+++ update/generate_database.py
@@ -19,10 +19,6 @@ def run(path: Path) -> None:
         database.update(pickle.load(f))
     logging.info("Finished integrating taxo")
 
-    with open(path / "database_biblio.pkl", "rb") as f:
-        database.update(pickle.load(f))
-    logging.info("Finished integrating biblio")
-
     with open(path / "database.pkl", "wb") as f:
         pickle.dump(database, f)
     logging.info("Finished dumping")